Rename app to 'export_academy'
