+++ config/settings.py
@@ -85,7 +85,7 @@ INSTALLED_APPS = [
     'health_check.cache',
     'sso_profile',
     'directory_components',
-    'events',
+    'export_academy',
 ]
 
 MIDDLEWARE = [
