--- events/apps.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-
-class EventsConfig(AppConfig):
-    name = 'events'
-
-    def ready(self):
-        from events import context  # noqa F401
