--- events/context.py
@@ -1,5 +1,0 @@
-from core.context import AbstractPageContextProvider
-
-
-class EventsContextProvider(AbstractPageContextProvider):
-    pass
