+++ export_academy/context.py
@@ -1,0 +1,5 @@
+from core.context import AbstractPageContextProvider
+
+
+class ExportAcademyContextProvider(AbstractPageContextProvider):
+    pass
